+++ src/makeSentence.py
@@ -1,4 +1,13 @@
 import markovify
+import json
+
+def remover(result):
+  # Load banned.json
+  json_open = open('data/banned.json', 'r')
+  json_load = json.load(json_open)
+  for w in json_load['words']:
+    result = result.replace(w, '')
+  return result
 
 def make_sentence():
     with open("data/model.json", "r") as f:
@@ -10,4 +19,4 @@ def make_sentence():
         if made:
             sentence = "".join(made.split())
             break
-    return sentence
+    return remover(sentence)
